requirements.txt
# ECHO OS Barebone - Dependencies

# Core FastAPI and server dependencies
fastapi>=0.115.0
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# HTTP and JSON handling
requests==2.31.0
httpx>=0.28.1

# Data validation
pydantic>=2.0.0

# Date and time handling
python-dateutil==2.8.2

# Text processing and encoding
charset-normalizer==3.3.2
unicodedata2==15.1.0

# Environment and configuration
python-dotenv==1.0.0

# AWS SDK for Layer 5 (DynamoDB, S3)
boto3==1.34.22

# OpenAI SDK for embeddings
openai>=1.55.3

# Anthropic SDK for Claude chat generation
anthropic==0.40.0

# Google Gemini SDK (optional)
google-genai==1.54.0

# Logging and monitoring
structlog==23.2.0

# Vector search
faiss-cpu
numpy
scipy

# JWT and authentication
PyJWT==2.8.0
passlib[bcrypt]>=1.7.4

# Rate limiting
slowapi==0.1.9

# Google Drive Integration (optional)
google-api-python-client==2.111.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Document Processing
pypdf==4.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Azure AI Search
azure-search-documents>=11.4.0
azure-core>=1.29.0
azure-identity>=1.15.0

# Microsoft Graph SDK
msgraph-core>=1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
